// integration/tristar/interfaces/tristar.enum.ts
export enum PersonType {
  PERSONAL = 1,
  COMPANY = 2,
}

export enum DocumentType {
  TAX_ID = 1,
  PASSPORT = 2,
}

export enum CountryID {
  BR = 32,
  CH = 46,
  PT = 179,
  US = 236,
}

export enum CountryCode {
  BR = 'BR',
  CH = 'CH',
  PT = 'PT',
  US = 'US',
}

export enum CurrencyID {
  USD = 1,
}

export enum SupervisoryOrgan {
  ANVISA = 1,
}

export enum ShipmentItem {
  PRODUCTS = 10,
  BOOKS = 20,
  MEDICINE = 30,
  CBD = 40,
  OTHER = 90,
}